BugID(s): VOIP-80387
merge win32 branch back into trunk 32569 trunk, head branch (32660-32837)

git-svn-id: https://subversion.polycom.com/SVN/RepoMisc/TestUtilities/trunk/SIPped/SIPped/src@32846 85c8ee13-dc6b-4ec3-b49e-93a6737f2991

+++ CompositeDocument.hpp
@@ -106,7 +106,7 @@ class CompositeDocument {
     string showStack();
     string topDoc();
     int getCompositeLineNumber();
-    int getDocsInStack();
+    size_t getDocsInStack();
     void reduceBothLineNumBy(int delta);
     // void incrCompositeLineNumber();
   }; // inner class DocStack
@@ -127,8 +127,8 @@ public:
   int getCompositeLineNumber();
   string strStackFromCompositeLineNumber(int compositeLineNumber);
   string getCurrDoc();
-  int currStackSize();
-  int getQtyStacks();
+  size_t currStackSize();
+  size_t getQtyStacks();
   string dumpStacks();
   int compositeLineNumberFromIndex(int index);
   string strStackFromIndex( int index);
